+++ Projects/Experimental/Toolkit/Vector/VectorPathToolInteractions.cs
@@ -1,4 +1,4 @@
-﻿//using Sachssoft.Sasogine.Input;
+//using Sachssoft.Sasogine.Input;
 
 //namespace Sachssoft.Sasogine.Experimental.Components.Tools
 //{